[P4ADEV-3159] refactor Events module structure and implement EventsContainer for better state management
