+++ src/routes/Events/configs.tsx
@@ -8,13 +8,16 @@ import {
   FilterItem
 } from '../../components/FilterContainer/FilterContainer';
 import { TabsConfig } from '../../components/SearchCard/SearchCard';
-import i118n from '../../translations/i18n';
 import { BaseFilterValues } from '../../models/Filters';
+import { ReadMore } from '@mui/icons-material';
+import i18n from '../../translations/i18n';
+import { GridColDef } from '@mui/x-data-grid';
+import { noFilterSetted } from '../../utils/filtersValidation';
 
 export const silFields: Array<FilterItem> = [
   {
     type: COMPONENT_TYPE.textField,
-    label: i118n.t('events.searchIUVDescription'),
+    label: i18n.t('events.searchIUVDescription'),
     adornment: <Search />,
     id: 'iuv'
   },
@@ -22,13 +25,13 @@ export const silFields: Array<FilterItem> = [
     type: COMPONENT_TYPE.dateRange,
     label: 'dateRange',
     required: true,
-    from: { label: i118n.t('events.searchDateFromDescription') },
-    to: { label: i118n.t('events.searchDateToDescription') },
+    from: { label: i18n.t('events.searchDateFromDescription') },
+    to: { label: i18n.t('events.searchDateToDescription') },
     id: 'eventDate'
   },
   {
     type: COMPONENT_TYPE.select,
-    label: i118n.t('events.searchEventDescription'),
+    label: i18n.t('events.searchEventDescription'),
     options: Object.entries(RegistrySilEventType).map(([key, value]) => ({
       label: key,
       value
@@ -40,7 +43,7 @@ export const silFields: Array<FilterItem> = [
 export const nodoFields: Array<FilterItem> = [
   {
     type: COMPONENT_TYPE.textField,
-    label: i118n.t('events.searchIUVDescription'),
+    label: i18n.t('events.searchIUVDescription'),
     adornment: <Search />,
     id: 'iuv'
   },
@@ -48,13 +51,13 @@ export const nodoFields: Array<FilterItem> = [
     type: COMPONENT_TYPE.dateRange,
     label: 'dateRange',
     required: true,
-    from: { label: i118n.t('events.searchDateFromDescription') },
-    to: { label: i118n.t('events.searchDateToDescription') },
+    from: { label: i18n.t('events.searchDateFromDescription') },
+    to: { label: i18n.t('events.searchDateToDescription') },
     id: 'eventDate'
   },
   {
     type: COMPONENT_TYPE.select,
-    label: i118n.t('events.searchEventDescription'),
+    label: i18n.t('events.searchEventDescription'),
     options: Object.entries(RegistryPagoPaEventType).map(([key, value]) => ({
       label: key,
       value
@@ -65,11 +68,11 @@ export const nodoFields: Array<FilterItem> = [
 
 export const tabs: Array<TabsConfig> = [
   {
-    label: i118n.t('events.tabs.sil'),
+    label: i18n.t('events.tabs.sil'),
     fields: silFields
   },
   {
-    label: i118n.t('events.tabs.nodo'),
+    label: i18n.t('events.tabs.nodo'),
     fields: nodoFields
   }
 ];
@@ -117,3 +120,80 @@ export function getQueryFromFilterValues<
     eventType: activeFilterValues.event as NodoOrSilEvent<T>['eventType']
   };
 }
+
+export const columns: Array<GridColDef> = [
+  {
+    field: 'eventDate',
+    headerName: i18n.t('event.DataEvento'),
+    flex: 1,
+    type: 'date'
+  },
+  {
+    field: 'org',
+    headerName: i18n.t('event.Ente'),
+    flex: 1,
+    type: 'string'
+  },
+  {
+    field: 'iuv',
+    headerName: i18n.t('event.IUV'),
+    flex: 1,
+    type: 'string'
+  },
+  {
+    field: 'event',
+    headerName: i18n.t('event.EVENTO'),
+    flex: 1,
+    type: 'string'
+  },
+  {
+    field: 'action',
+    headerName: '',
+    flex: 0.5,
+    sortable: false,
+    align: 'right',
+    headerAlign: 'right',
+    renderCell: () => (
+      <div
+        style={{
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'flex-end',
+          height: '100%',
+          width: '100%'
+        }}
+      >
+        <ReadMore color="primary" onClick={console.log} />
+      </div>
+    )
+  }
+];
+
+export const getFilters = (
+  registryType: RegistryType,
+  onSubmit: () => void
+) => {
+  const fields = registryType === 'pagopa' ? nodoFields : silFields;
+  const filters = [
+    ...fields.map((el) => ({ ...el, gridWidth: 3 })),
+    {
+      type: COMPONENT_TYPE.button,
+      label: 'cerca',
+      gridWidth: 3,
+      onClick: onSubmit
+    }
+  ];
+
+  return filters;
+};
+
+export type RegistryType = 'pagopa' | 'sil';
+
+// Create a deep copy of filterValues to avoid mutating the state directly
+export function deepCopy<T>(obj: T) {
+  return structuredClone<T>(obj);
+}
+
+export const testFilterValidity = (
+  filterValues: NodoFilterValues | SilFilterValues
+) => !noFilterSetted(filterValues);
